+++ src/Backends/Backend.php
@@ -23,7 +23,7 @@ interface Backend extends Stringable
      * @param callable(mixed,mixed):void $after
      * @param mixed $context
      */
-    public function enqueue(Task $task, ?callable $after = null, $context = null) : void;
+    public function enqueue(Task $task, ?callable $after = null, mixed $context = null) : void;
 
     /**
      * Process the queue and return the results.