+++ src/Datasets/Dataset.php
@@ -197,6 +197,21 @@ abstract class Dataset implements ArrayAccess, IteratorAggregate, Countable
     }
 
     /**
+     * Drop a feature column at a given offset from the dataset.
+     *
+     * @param int $offset
+     * @return self
+     */
+    public function dropFeature(int $offset) : self
+    {
+        foreach ($this->samples as &$sample) {
+            array_splice($sample, $offset, 1);
+        }
+
+        return $this;
+    }
+
+    /**
      * Rotate the sample matrix so that the values of each feature become rows.
      *
      * @return mixed[]
@@ -575,10 +590,10 @@ abstract class Dataset implements ArrayAccess, IteratorAggregate, Countable
      * @internal
      *
      * @param int $offset
-     * @param mixed $value
+     * @param string|int|float $value
      * @return array{self,self}
      */
-    abstract public function splitByFeature(int $offset, $value) : array;
+    abstract public function splitByFeature(int $offset, string|int|float $value) : array;
 
     /**
      * Partition the dataset into left and right subsets based on the samples' distances from two centroids.