+++ src/Graph/Nodes/Split.php
@@ -30,9 +30,9 @@ class Split implements Decision, HasBinaryChildren
     /**
      * The value to split on.
      *
-     * @var int|float|string
+     * @var string|int|float
      */
-    protected $value;
+    protected string|int|float $value;
 
     /**
      * The left and right subsets of the training data.
@@ -63,7 +63,7 @@ class Split implements Decision, HasBinaryChildren
      * @param int<0,max> $n
      * @throws \Rubix\ML\Exceptions\InvalidArgumentException
      */
-    public function __construct(int $column, $value, array $subsets, float $impurity, int $n)
+    public function __construct(int $column, string|int|float $value, array $subsets, float $impurity, int $n)
     {
         $this->column = $column;
         $this->value = $value;
@@ -85,9 +85,9 @@ class Split implements Decision, HasBinaryChildren
     /**
      * Return the split value.
      *
-     * @return int|float|string
+     * @return string|int|float
      */
-    public function value()
+    public function value() : string|int|float
     {
         return $this->value;
     }