+++ src/Loggers/Screen.php
@@ -2,6 +2,8 @@
 
 namespace Rubix\ML\Loggers;
 
+use Stringable;
+
 use function trim;
 use function date;
 use function strtoupper;
@@ -45,10 +47,10 @@ class Screen extends Logger
      * Logs with an arbitrary level.
      *
      * @param mixed $level
-     * @param string $message
+     * @param string|Stringable $message
      * @param mixed[] $context
      */
-    public function log($level, $message, array $context = []) : void
+    public function log($level, string|Stringable $message, array $context = []) : void
     {
         $prefix = '';
 
@@ -62,6 +64,6 @@ class Screen extends Logger
 
         $prefix .= strtoupper((string) $level);
 
-        echo $prefix . ': ' . trim($message) . PHP_EOL;
+        echo $prefix . ': ' . trim((string) $message) . PHP_EOL;
     }
 }