+++ src/NeuralNet/Layers/Hidden.php
@@ -4,7 +4,6 @@ namespace Rubix\ML\NeuralNet\Layers;
 
 use Rubix\ML\Deferred;
 use Rubix\ML\NeuralNet\Optimizers\Optimizer;
-use Stringable;
 
 /**
  * Hidden
@@ -13,7 +12,7 @@ use Stringable;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-interface Hidden extends Layer, Stringable
+interface Hidden extends Layer
 {
     /**
      * Calculate the gradient and update the parameters of the layer.